Split base huffmantable stuff into AbstractHuffmanTable.
