+++ src/librawspeed/decompressors/CMakeLists.txt
@@ -2,6 +2,7 @@ FILE(GLOB SOURCES
   "AbstractDecompressor.h"
   "AbstractDngDecompressor.cpp"
   "AbstractDngDecompressor.h"
+  "AbstractHuffmanTable.h"
   "AbstractLJpegDecompressor.cpp"
   "AbstractLJpegDecompressor.h"
   "AbstractParallelizedDecompressor.cpp"
